requirements.txt
torch
torchvision
matplotlib
Pillow
scikit-learn
numpy
